Extract as the file
