+++ lib/fluent/plugin/in_tail.rb
@@ -22,6 +22,7 @@ require 'fluent/event'
 require 'fluent/plugin/buffer'
 require 'fluent/plugin/parser_multiline'
 require 'fluent/variable_store'
+require 'fluent/plugin/in_tail/position_file'
 
 if Fluent.windows?
   require_relative 'file_wrapper'
@@ -908,140 +909,6 @@ module Fluent::Plugin
         def reset_timer
           @start = Time.now
         end
-      end
-    end
-
-    class PositionFile
-      UNWATCHED_POSITION = 0xffffffffffffffff
-
-      def initialize(file, file_mutex, map, last_pos)
-        @file = file
-        @file_mutex = file_mutex
-        @map = map
-        @last_pos = last_pos
-      end
-
-      def [](path)
-        if m = @map[path]
-          return m
-        end
-
-        @file_mutex.synchronize {
-          @file.pos = @last_pos
-          @file.write "#{path}\t0000000000000000\t0000000000000000\n"
-          seek = @last_pos + path.bytesize + 1
-          @last_pos = @file.pos
-          @map[path] = FilePositionEntry.new(@file, @file_mutex, seek, 0, 0)
-        }
-      end
-
-      def self.parse(file)
-        compact(file)
-
-        file_mutex = Mutex.new
-        map = {}
-        file.pos = 0
-        file.each_line {|line|
-          m = /^([^\t]+)\t([0-9a-fA-F]+)\t([0-9a-fA-F]+)/.match(line)
-          unless m
-            $log.warn "Unparsable line in pos_file: #{line}"
-            next
-          end
-          path = m[1]
-          pos = m[2].to_i(16)
-          ino = m[3].to_i(16)
-          seek = file.pos - line.bytesize + path.bytesize + 1
-          map[path] = FilePositionEntry.new(file, file_mutex, seek, pos, ino)
-        }
-        new(file, file_mutex, map, file.pos)
-      end
-
-      # Clean up unwatched file entries
-      def self.compact(file)
-        file.pos = 0
-        existent_entries = file.each_line.map { |line|
-          m = /^([^\t]+)\t([0-9a-fA-F]+)\t([0-9a-fA-F]+)/.match(line)
-          unless m
-            $log.warn "Unparsable line in pos_file: #{line}"
-            next
-          end
-          path = m[1]
-          pos = m[2].to_i(16)
-          ino = m[3].to_i(16)
-          # 32bit inode converted to 64bit at this phase
-          pos == UNWATCHED_POSITION ? nil : ("%s\t%016x\t%016x\n" % [path, pos, ino])
-        }.compact
-
-        file.pos = 0
-        file.truncate(0)
-        file.write(existent_entries.join)
-      end
-    end
-
-    # pos               inode
-    # ffffffffffffffff\tffffffffffffffff\n
-    class FilePositionEntry
-      POS_SIZE = 16
-      INO_OFFSET = 17
-      INO_SIZE = 16
-      LN_OFFSET = 33
-      SIZE = 34
-
-      def initialize(file, file_mutex, seek, pos, inode)
-        @file = file
-        @file_mutex = file_mutex
-        @seek = seek
-        @pos = pos
-        @inode = inode
-      end
-
-      def update(ino, pos)
-        @file_mutex.synchronize {
-          @file.pos = @seek
-          @file.write "%016x\t%016x" % [pos, ino]
-        }
-        @pos = pos
-        @inode = ino
-      end
-
-      def update_pos(pos)
-        @file_mutex.synchronize {
-          @file.pos = @seek
-          @file.write "%016x" % pos
-        }
-        @pos = pos
-      end
-
-      def read_inode
-        @inode
-      end
-
-      def read_pos
-        @pos
-      end
-    end
-
-    class MemoryPositionEntry
-      def initialize
-        @pos = 0
-        @inode = 0
-      end
-
-      def update(ino, pos)
-        @inode = ino
-        @pos = pos
-      end
-
-      def update_pos(pos)
-        @pos = pos
-      end
-
-      def read_pos
-        @pos
-      end
-
-      def read_inode
-        @inode
       end
     end
   end
